Renamed `plugins` folder to `features`
